+++ src/ui/log_tab.rs
@@ -373,6 +373,10 @@ impl Tab for LogTab<'_> {
         self.details_keybinds = DetailsPanelKeybinds::new();
     }
 
+    fn toggle_layout(&mut self) {
+        self.pane_divider.toggle_layout();
+    }
+
     fn is_stale(&self) -> bool {
         self.stale
     }